requirements.txt
aider-chat==0.57.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anthropic==0.31.2
anyio==4.6.0
arrow==1.3.0
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
deprecation==2.1.0
diff-match-patch==20230430
diskcache==5.6.3
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
flake8==7.1.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
gotrue==2.6.0
greenlet==3.0.3
grep-ast==0.3.3
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==7.2.1
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
litellm==1.47.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
multidict==6.1.0
networkx==3.2.1
numpy==1.26.4
openai==1.47.0
packaging==24.1
pathspec==0.12.1
pexpect==4.9.0
pika==1.2.0
pillow==10.4.0
postgrest==0.16.9
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
pypandoc==1.13
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
realtime==1.0.6
redis==5.0.7
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
rq==1.16.2
scipy==1.13.1
sentry-sdk==2.14.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.0
soundfile==0.12.1
soupsieve==2.6
SQLAlchemy==2.0.31
storage3==0.7.7
StrEnum==0.4.15
supabase==2.5.3
supafunc==0.4.7
tiktoken==0.7.0
times==0.7
tokenizers==0.19.1
tqdm==4.66.5
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
yarl==1.11.1
zipp==3.20.2
